requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.40
WTForms==3.1.1
Werkzeug>=3.1
reportlab==4.0.8
weasyprint==60.2
openpyxl==3.1.2
gunicorn==21.2.0
python-dotenv==1.0.0
bcrypt==4.1.2

